Add a test helper for dealing with the test token.
This test token helper will be re-used in the terraform provider and was
extracted from it. It allows a file on disk, or in the environment to
manipulate the token value in an easy to consume form.

This also adds another function to gate the test runs on the token's
existence; so build tags aren't required to run a full suite. Just the
tests that are not flagged will run.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>