release tooling
This creates a VERSION file at the root and some go:generate trampolines
in the make system to set the versioning properly in the source, as well
as the repostiory. You can create a new release by typing:

VERSION=<version, 'v' is not added> make release

Which will generate the tag, version.go and commit the result, ready to
push.

Signed-off-by: Erik Hollensbe <github@hollensbe.org>