Basic client rate limiting
This grabs these response headers and
caches them.

X-Ratelimit-Limit: 20
X-Ratelimit-Remaining: 14
X-Ratelimit-Reset: Tue, 06 Aug 2024 20:53:48 UTC

The we use that to pause requests so we don't
go over the limits and get 429 errors.

The tests pass.

If someone knows a better wait algorithm, go for it.

Right now it's just:
(Limit - Remaining) * 10 milliseconds
20 - 19 -> 10ms
20 - 2 -> 180ms